+++ zebrad/src/commands.rs
@@ -10,13 +10,17 @@ use crate::config::ZebradConfig;
 
 pub use self::{entry_point::EntryPoint, start::StartCmd};
 
-use self::{copy_state::CopyStateCmd, generate::GenerateCmd, tip_height::TipHeightCmd};
+use self::{
+    copy_state::CopyStateCmd, generate::GenerateCmd, rollback_state::RollbackStateCmd,
+    tip_height::TipHeightCmd,
+};
 
 pub mod start;
 
 mod copy_state;
 mod entry_point;
 mod generate;
+pub mod rollback_state;
 mod tip_height;
 
 #[cfg(test)]
@@ -36,6 +40,9 @@ pub enum ZebradCmd {
 
     /// Generate a default `zebrad.toml` configuration
     Generate(GenerateCmd),
+
+    /// Roll back Zebra's finalized chain state on disk
+    RollbackState(RollbackStateCmd),
 
     /// Start the application (default command)
     Start(StartCmd),
@@ -57,7 +64,7 @@ impl ZebradCmd {
             CopyState(_) | Start(_) => true,
 
             // Utility commands that don't use server components
-            Generate(_) | TipHeight(_) => false,
+            Generate(_) | RollbackState(_) | TipHeight(_) => false,
         }
     }
 
@@ -71,7 +78,7 @@ impl ZebradCmd {
             Start(_) => true,
 
             // Utility commands
-            CopyState(_) | Generate(_) | TipHeight(_) => false,
+            CopyState(_) | Generate(_) | RollbackState(_) | TipHeight(_) => false,
         }
     }
 
@@ -90,7 +97,7 @@ impl ZebradCmd {
             // This output:
             // - is used by automated tools, or
             // - needs to be read easily.
-            Generate(_) | TipHeight(_) => true,
+            Generate(_) | RollbackState(_) | TipHeight(_) => true,
 
             // Commands that generate informative logging output by default.
             CopyState(_) | Start(_) => false,
@@ -111,6 +118,7 @@ impl Runnable for ZebradCmd {
         match self {
             CopyState(cmd) => cmd.run(),
             Generate(cmd) => cmd.run(),
+            RollbackState(cmd) => cmd.run(),
             Start(cmd) => cmd.run(),
             TipHeight(cmd) => cmd.run(),
         }